+++ backend-django/app/eric/drives/admin.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 #
 """ contains basic admin functionality for eric workbench pictures """
-
+from admin_auto_filters.filters import AutocompleteFilter
 from django.contrib import admin
 from django.contrib.admin import TabularInline
 from django.contrib.auth import get_user_model
@@ -21,6 +21,26 @@ class DirectoryInline(TabularInline):
     verbose_name = _("Directory")
     verbose_name_plural = _("Directories")
     extra = 1
+
+
+class DriveFilter(AutocompleteFilter):
+    title = 'Drive'
+    field_name = 'drive'
+
+
+@admin.register(Directory)
+class DirectoryAdmin(admin.ModelAdmin):
+    list_display = (
+        'name',
+        'directory',
+        'drive',
+    )
+    list_filter = (
+        DriveFilter,
+    )
+    search_fields = (
+        'title',
+    )
 
 
 @admin.register(Drive)