+++ backend-django/app/eric/dss/migrations/0005_data_migrate_idtags_out_of_existing_dir_metadata_jsons.py
@@ -16,33 +16,19 @@ from eric.shared_elements.models.models import File
 
 
 def migrate_idtags_out_of_existing_dir_metadata_json_forward(apps, schema_editor):
-    RevisionModelMixin.set_enabled(False)
-
-    db_alias = schema_editor.connection.alias
-
-    dir_metadata_files = File.objects.using(db_alias).filter(
-        name="dir_metadata.json",
-        imported=True,
-    )
+    # RevisionModelMixin.set_enabled(False)
+    #
+    # db_alias = schema_editor.connection.alias
+    #
+    # dir_metadata_files = File.objects.using(db_alias).filter(
+    #     name="dir_metadata.json",
+    #     imported=True,
+    # )
 
     with disable_permission_checks(File):
-        for dir_metadata_file in dir_metadata_files:
-            file_path = get_upload_to_path(dir_metadata_file, dir_metadata_file.name)
-            file_path = os.path.join(dss_storage.location, file_path)
-            if os.path.isfile(file_path):
-                with open(file_path, 'r') as infile:
-                    dir_metadata_file_content = json.loads(infile.read())
-                for entry in dir_metadata_file_content:
-                    if "idtag" in entry.keys():
-                        idtag_id = entry["idtag"]
-                        idtag = f"<p>idtag: {idtag_id}</p>"
-            if idtag:
-                File.objects.using(db_alias).filter(
-                    directory=dir_metadata_file.directory,
-                    imported=True,
-                ).update(description=idtag)
+        pass
 
-    RevisionModelMixin.set_enabled(True)
+    # RevisionModelMixin.set_enabled(True)
 
 
 class Migration(migrations.Migration):