+++ frontend-angular/apps/eworkbench/src/app/modules/shared/filter-sidebar/filter-sidebar.component.ts
@@ -65,6 +65,7 @@ export class FilterSidebarComponent implements OnInit {
     this.user$.pipe(take(1)).subscribe(user => {
       const open = user.userprofile.ui_settings?.filter_sidebar?.open;
       this.open$.next(open ?? false);
+      this.toggle.next(open ?? false);
     });
   }